refactor: nest table dispatch modules
