+++ crates/evm2/src/interpreter/dispatch/mod.rs
@@ -2,14 +2,15 @@
 
 #[cfg(tco)]
 use crate::interpreter::gas::RemainingGas;
+#[cfg(tco)]
+use crate::interpreter::{Interpreter, Stack};
+#[cfg(tco)]
+use crate::trustme;
 use crate::{
     BaseEvmConfigSelector, EvmConfig, EvmConfigSelector, EvmTypes, VersionTables,
     evm::config::SelectorVersionTables,
-    interpreter::{InstrStop, Interpreter, InterpreterState, Pc, Result, Stack, StackMut, op},
-    trustme,
+    interpreter::{InstrStop, InterpreterState, Pc, Result, StackMut, op},
 };
-#[cfg(not(tco))]
-use core::hint::cold_path;
 
 #[cold]
 pub(crate) const fn unknown_instruction<T: EvmTypes>(
@@ -26,47 +27,11 @@ cfg_if::cfg_if! {
     if #[cfg(tco)] {
         mod tco;
         use tco as imp;
-    } else if #[cfg(dispatch_packed)] {
-        mod packed;
-        use packed as imp;
-    } else if #[cfg(dispatch_single_return)] {
-        mod single_return;
-        use single_return as imp;
     } else {
-        mod unpacked;
-        use unpacked as imp;
-    }
-}
+        mod table;
+        use table as imp;
 
-#[cfg(not(tco))]
-pub(in crate::interpreter) fn run_table_loop<T: EvmTypes>(
-    interpreter: &mut Interpreter<'_, T>,
-    instructions: &InstrTable<T>,
-) -> InstrStop {
-    // SAFETY: Only the active interpreter lifetime is erased; this stays as a raw pointer so
-    // the dispatch loop does not create an extra `&mut` alias for `interpreter`.
-    let raw = unsafe { trustme::decouple_lt_mut_ptr(interpreter as *mut Interpreter<'_, T>) };
-    // SAFETY: Instruction methods must not access the stack through `InterpreterState` while
-    // the separate stack view is live.
-    let state = InterpreterState::wrap_mut(unsafe { &mut *raw });
-    let mut pc = Pc::new(interpreter.pc);
-    let mut stack = Stack::new(&mut interpreter.stack, interpreter.stack_len);
-    let mut loop_state = imp::loop_state(&interpreter.gas);
-    loop {
-        let op = pc.op();
-        let instr = instructions[op as usize];
-        let (next_pc, next_stack_len) =
-            imp::dispatch_loop_call(instr, pc, stack.reborrow(), state, &mut loop_state);
-        pc = next_pc;
-        stack.len = next_stack_len;
-
-        if pc.as_ptr().is_null() {
-            cold_path();
-            interpreter.pc = pc.as_ptr();
-            interpreter.stack_len = stack.len;
-            imp::finish_loop(&mut interpreter.gas, loop_state);
-            return interpreter.result.unwrap_err();
-        }
+        pub(in crate::interpreter) use table::run_table_loop;
     }
 }
