+++ crates/evm2/src/interpreter/dispatch/table/single_return.rs
@@ -6,7 +6,7 @@ use crate::{
 use core::hint::cold_path;
 
 /// Single-return instruction function pointer.
-pub(super) type RawInstrFn<T> = extern_table!(
+pub(in crate::interpreter::dispatch) type RawInstrFn<T> = extern_table!(
     fn(
         pc: Pc,
         stack: Stack<'_>,
@@ -16,13 +16,13 @@ pub(super) type RawInstrFn<T> = extern_table!(
 );
 
 /// Single-return instruction dispatch table.
-pub(super) type RawInstrTable<T> = [RawInstrFn<T>; 256];
+pub(in crate::interpreter::dispatch) type RawInstrTable<T> = [RawInstrFn<T>; 256];
 
 #[inline(always)]
-pub(super) const fn loop_state(_gas: &Gas) {}
+pub(in crate::interpreter::dispatch) const fn loop_state(_gas: &Gas) {}
 
 #[inline(always)]
-pub(super) fn dispatch_loop_call<T: EvmTypes>(
+pub(in crate::interpreter::dispatch) fn dispatch_loop_call<T: EvmTypes>(
     instr: RawInstrFn<T>,
     pc: Pc,
     stack: Stack<'_>,
@@ -35,10 +35,10 @@ pub(super) fn dispatch_loop_call<T: EvmTypes>(
 }
 
 #[inline(always)]
-pub(super) const fn finish_loop(_gas: &mut Gas, _loop_state: ()) {}
+pub(in crate::interpreter::dispatch) const fn finish_loop(_gas: &mut Gas, _loop_state: ()) {}
 
 extern_table! {
-    pub(super) fn dispatch<
+    pub(in crate::interpreter::dispatch) fn dispatch<
         T: EvmTypes,
         C: EvmConfig<T>,
         M: InspectMode<T>,
@@ -60,7 +60,7 @@ extern_table! {
         )
     }
 
-    pub(super) fn unknown_dispatch<
+    pub(in crate::interpreter::dispatch) fn unknown_dispatch<
         T: EvmTypes,
         C: EvmConfig<T>,
         M: InspectMode<T>,
